fix: move versions check to init function and refactor store
